add notification

+++ database/migrations/2022_06_03_023624_create_surat_permohonan_bebas_lab_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('surat_permohonan_bebas_lab', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_mahasiswa_id')->constrained('user_mahasiswa');
-            $table->int('no_surat');
+            $table->bigInteger('no_surat');
             $table->string('keterangan');
             $table->text('judul');
             $table->foreignId('status_id')->constrained('status_aktivasi');
